refactor: :recycle: move to parent directory
Logo, MenuIcon -> /components/
